MERO-2191: Fixed assert(flags & M0_NET_BUF_QUEUED) at m0_net_buffer_event_post()
In case that more than one event is needed for a buffer to receive and network
error is occured before the transmission is finished and lnet signals that
buffer is still in use and should not be dequeued, our lnet code missed to
set M0_NET_BUF_RETAIN flag. This led to dequeue of the buffer that is still
in use and crash in m0_net_buffer_event_post() on the next call, when another
part of the buffer was successfully received after the error.

Change-Id: Id439dce80d9cafc57a96d7f26efd5b0f31ebc764