MERO-2297: rpm: exclude source code from -debuginfo package
By default rpmbuild creates a *-debuginfo package which contains
debugging data along with source code. That's unfortunate, because mero
rpms can be distributed to third parties. And not all of them are
supposed to have access to the source code. So we need to prevent it
from getting into mero-debuginfo.rpm.

Still, it's possible to build debuginfo rpm with sources using special
rpm flag:

  make rpms RPMBUILD_FLAGS='--with debuginfo_sources'

Change-Id: Ie2496840643164d03c271b279b14083db7d0df8e