[MERO-2303] ha/ut: Fix `ha-state-ut:ha-poolversion-get' UT
Sequence of events leading to the intermittent failure of
test_poolversion_get() a.k.a. the bug MERO-2303:

0. ha_ut_conf_init() calls m0_ha_state_set(), which initiates asynchronous
   modification of m0_conf_obj::co_ha_state of disk objects --- when the
   asynchronous operation completes, .co_ha_state will be set to
   M0_NC_ONLINE.

1. m0_ha_state_accept() synchronously modifies .co_ha_state of a disk object,
   setting it to non-M0_NC_ONLINE.

2. m0_conf_pver_get() reaches conf_pver_recd_build(), which reads
   .co_ha_state of disk objects and fills recd[] vector.

  test_poolversion_get
   \_ ha_ut_conf_init
   |   \_ m0_ha_state_set
   |       \_ m0_ha_msg_nvec_send
   |           \_ m0_ha_link_send  --  [0] initiates .co_ha_state modification
   \_ m0_ha_state_accept
   |   \_ m0_ha_clients_iterate
   |       \_ iter
   |           \_ ha_state_accept  --  [1] sets .co_ha_state
   \_ m0_conf_pver_get
       \_ m0_conf_pver_find
           \_ conf_pver_find_locked
               \_ conf_pver_base_recd_update
                   \_ m0_conf_walk
                       \_ conf_walker_down
                           \_ conf_pver_recd_build  --  [2] reads .co_ha_state

test_poolversion_get() did not consider the fact that the setting of disk
objects' .co_ha_status to M0_NC_ONLINE, initiated by m0_ha_state_set() call,
may occur at any moment, as it is executed in a different thread.

The test succeeded, because the operation, started by [0], tended to
complete before [1].  But in some cases it completed _after_ [1], the
non-M0_NC_ONLINE states got overwritten, and test's expectations were
broken.

The fix is not to call m0_ha_state_set() in ha_ut_conf_init(). It's not
needed there anyway.

Change-Id: I58033dd5c8af6cb8fb7c55dd38e8b34c0342dce3