MERO-2398: Mero panic: (!m0_tlink_is_in(d, obj)) at m0_tlist_add_tail()
In case of a failure sns operation is aborted, during this alive nodes
detect frozen aggregation groups and destroy them. Due to this there
could be some onwire incoming copy packets which may be received at
the receiver end but the aggregation group may have been already marked
frozen (due to failure). In such a case sns replies with -EINVAL due
to which the sender fails the copy machine and thus the operation.

Fix:
- Instead of returning -EINVAL return -ENOENT, in which case sender
  will not fail the copy machine but abort cleanly. SNS ignores
  -ENOENT and continues further.

Change-Id: I0789ebfd7fee78e17ed115a00c582df9bfd9b9ea