MERO-2406: M0_ASSERT_INFO(stob_state == CSS_EXISTS) through cp_io()
Problem:
During concurrent file delete with sns operation (repair/rebalance),
it is possible that a stob is deleted and moved to idle list in stob
cache but sns operation copy packet can try to find the stob again
during storage io. Especially in case of sns read a deleted stob
can be found in the cache with stob_state as CSS_NOENT. In case of
read sns operation a found stob is not initialised. This leads to
above assertion.

Fix:
Instead of asserting the stob_state to CSS_EXISTS, check if stob_state
is CSS_NOENT and return -ENOENT to drop the copy packet.

Change-Id: Ice8d32ba8740057ff82f99ee85b679fa0a81b424