spiel: add a command to load a library in the process address space.
Add a new spiel function m0_spiel_process_lib_load() that loads a
library, specified by the pathname in the server file system, into the
process address space.

Library loading is supported only in user space at the moment. The
library must be loadable by dlopen(3).

When the library is loaded, mero_lib_init() function in it (if
present) is invoked without parameters. This function is called in a
fom tick context, so it should not block.

Also, add full build directory pathname to m0_build_info.

Change-Id: Id1b5374f64ddcd679dee7b014227f2159b658c48