MERO-2400 getattr crashes due to absence of pool version in mdcob
When unlink fails, due to one of the device being detached, mdcobs are
deleted, but inode continues to exist. Then if a new io is done on
the same file, setattr creates mdcob without the pool version attribute.
Due to this next gettattr fails as pool version is absent.

This is fixed by setting the pool version present in the inode to mdcob
as an attribute during setattr operation.

Change-Id: I1659e9c5e0ab150066280f53d322e6990442e6b7