[MERO-2436] conf (m0_conf_pver_find): Skip imeta pver
Modify m0_conf_pver_find() to not return the "imeta" pver.
Disk in the imeta pver has path of /dev/null, because CAS doesn't actually
write to it (it uses BE). If IOS tries to write to the imeta pver, it will
crash.

Use zeroed value (M0_FID0) of imeta pver in unit and system tests.

Reviewed-by: Maxim Medved <max.medved@seagate.com>
Change-Id: I452639f6410d8a94e76d0e4062ebc9e5a6d1edea