MERO-2460 Do not update a failure for REPAIRING to FAILED transition
After repair is aborted due to second failure, it is seen that number of
device failures in a pool machine is increased again for first disk.
This is due to first disk is transitioned back to FAILED and "pst_nr_failures"
is increased in this case too.

This is fixed by not incrementing the failures when REPAIRING to FAILED
transition happens.

Also made a change to not allow REBALANCING to TRANSIENT or FAILED
transition, in such case disk will be moved to REPAIRED.

Change-Id: Iae26a9df1378e357bccbb766ee9d1be0e3c7d0c3