MERO-2455 be/log_store: add lsc_stob_dont_zero to configuration
If the option is enabled log_store doesn't zero backing stob on
creation. It reduces amount of I/O, but should be used only if
the stob is zeroed in other way.

Enable lsc_stob_dont_zero for all UTs. Running time of the be-ut
has been reduced by 1/3 on system with HDD.

Change-Id: Id36a1b9dfe2925fa94d6563dd19eec1aebc0356c