MERO-2470 m0d: Read conf of current process' services only
- cs_conf_services_init(): Take only services of current process into
consideration. Do this by process fid.
- Update affected UT/ST suites to comply with the new approach.

The present situation when the question whether service is local or not
is answered based the fact of service endpoint matching to rpc machine
endpoint looks not quite normal. This leads to multiple cases in tests
when configuration contains services with identical endpoints but
subordinating to different processes. In its turn this causes memory
leaks during service registrations.

Change-Id: Ie0aa297b43dcc4568e316f6b82d80020897f206c