MERO-2467: sns aggregation groups not cleaned up in case of m0d failure and restart
Problem:
On ioservice failure during sns repair, copy machines using proxy protocol
understand the failure and try to abort the operation implicitly.
A situation can be created where an ioservice has incoming aggregation
groups, expecting copy packets from the failed ioservice but (which
would never come) is unable to cleanup. Cleanup of frozen aggregation
groups happens in context of a sliding window update from remote node
corresponding to a proxy or when a copy machine sends a sliding window
update to a proxy. sns does not send update to a failed proxy in this
case the cleanup of frozen aggregation groups on the failed proxy
didn't complete. Also with pump fom waiting for completion signal,
sliding window update path is not exercised which also doesn’t help in
triggering cleanup.

Fix:
- Trigger a cleanup through sliding window update path in case of failed
  proxy without actually sending any update.
- Unset pool version dirty flag if number of failures are within allowed
  limits.

Change-Id: I14117fcb4f30ce440b83b2ed8a126df2a6aba216