MERO-2448: pool: add conf_ready clink when conf_expired happen
Pool common is interested in conf_ready event only when conf_expired event
happened. Therefore, no need to add clink to the chan in m0_pool_common_init().
Otherwise, this callback may be triggered on rconfc start and that leads to the
problem when reqh service context tries to make second subscription to the same
chan.

Change-Id: I73a49d0d0f23f3fef68f5a3c1d930f71179f15cd