MERO-1832 : After aborting rebalance when repair is restarted an assert (ag != NULL) is hit.
Added a case in m0t1fs_spiel_sns_repair.sh ST for handling the case
where the device currently being rebalanced fails. In this case, the device
the rebalance would be aborted, the device would be moved to "REPAIRED" state
and rebalance would be triggered again.

In case of a new device failure during rebalance, the device(s) being
rebalanced will be put to REPAIRED state and the repair would be
re-triggered.

Change-Id: I885c39ae5a11915f0504f5176c73400325add0ba