MERO-1718: be/alloc: added free chunks hash table
Simple and small hash table was added for the free chunks
to quickly reuse them on new allocations without splitting
(and merging). Here we utilise the fact that we have only
small set of different size allocations from BE segment.

That fixes performance issues when some files are deleted
(and, as result, many allocator chunks are freed) and
decreases the segment usage footprint compared with another
approach when all the freed chunks are added to the tail of
the free list and new allocations are made from the last
Big chunk only (refer to commit 0a4675c). The mentioned
approach also does not help in the long term when all the
segment space is used and the last Big chunk becomes an
ordinary small one.

Zones are switched OFF for now to align with the format
of BE in 1.0 (Teacake) until proper SW update is ready.

Reviewed-by: Maxim Medved <max.medved@seagate.com>

Change-Id: Iba25fd38476a71510dd857dbe291eb550a8558b3