clovis: read object which was written with old pool version
Consider an object obj1. When object was created and data was written
onto it, a pool version, it was associated was PV1.
PV1 consists of disks, d1, d2, d3, d4. Object was striped over d1 and
d3.
Now, disk d3 fails, SNS repair will start. Meanwhile, confd will create
a new pool version, PV2 containing d1, d2, and d4.
Now, read operation is done on obj1, as clovis does not know the pool
version associated with this object and as it always picked the latest
pool version, it will try to fetch data from the disks present in PV2.
As, disk d3 is not present in PV2, data will not be read properly.

Solution:
m0_clovis_obj_init() will contact service side for the pool version of
an object and will cache it in obj structure.

Change-Id: Icf5f9af304758306f9fba7fea3af28b2384464e0
Jira: Mero-2471