MERO-2510: Decrease buffer reservation count on buffer release instead of buffer attach
Issue:
Buffers are reserved for incoming copy packets on receiver copy machine.
This reservation count is decremented currently at the receiver once
buffer is consumed, but not yet released. While checking for available
space for more incoming copy packets, we check the sum of reserved number
of buffers and required number of buffers, if the count is less than
available number of free buffers then new aggregation groups are allocated
and sliding window is updated.
Here, because the reservation count is decremented (leads sns to think
that it can accommodate more incoming copy packets) early i.e. on consumption,
even before the actual buffers are released, leads to this problem as new
aggregation groups are allocated and sliding window is updated, so the sender
sends more copy packets which in turn do not find available buffers on receiver
end.This may happen intermittently if new aggregation groups are allocated
before the buffers are actually available for consumption..

Fix:
- remove a buffer from reservation after the buffer is consumed and
  released.
- Handle cleanup on proxy failure event properly.

st/m0t1fs_test.sh: small regression
Author: Andriy Tkachuk <andriy.tkachuk@seagate.com>
Reviwed at: http://es-gerrit.xyus.xyratex.com:8080/#/c/15524/

In prev. commit 136e21f there was dummy mistake in the new code
which did not allow the test to pass.

Change-Id: If339316eb3a3438890778e4609e0d43dcf4fb226