MERO-2127: decrease a chance of trace buffer overflow
Trace buffer is used as a ring-buffer and can be accessed by multiple
threads in parallel and concurrently. Each thread reserves a chunk of
the buffer and copies a trace record content there. Due to a
non-blocking design, there is a possibility of one thread overwriting a
whole buffer content while other thread(s) may be preempted in a midst
of copying trace record into the buffer. It can happen more or less
likely depending on the buffer size, number of CPU cores, number of
threads and a rate in which particular M0_LOG() entry is executed at run
time. In any case, some trace records in the buffer may end up corrupted.

To decrease probability of such trace buffer overflow the default size
of a trace buffer is increased to 16MiB (the old value was 4MiB). This
new value is purely empirical and it's been established by trial and
error in several different environments.

Also, M0_LOG() entry in m0_format_footer_verify_generic() is removed,
which could produce lots of trace records under some sns repair loads.

The code from 'lib/trace.c' which copies strings into trace buffer is
updated to use local copy of trace record to avoid SIGSEGV in case of
trace buffer overflow.

Change-Id: Ib93ca1ff899673ef403a7ee9971d0d931a30689e