MERO-2519 cas/service: fix checking of slave fom rc (CROW)
Problem:
simultaneous put of keys with CROW flag into empty index
may lead to fail due to index creation race.

Example:
simultaneous put of keys key1 and key2 into the same
index index1, index1 is empty on this node and is not yet
phisically created as CROW is used for distributed indices,
keys are sent in a separate fops fop1 and fop2:

1.1. put fop1 recevied -> fom1
1.2. put fop2 recevied -> fom2
2.1. fom1 -> lookup index1 -> ENOENT
2.2. fom2 -> lookup index1 -> ENOENT
3.1. fom1 -> run fom1.1 to create index1
3.2. fom2 -> run fom2.1 to create index1
4.1. fom1.1 -> create index1 -> SUCCESS
4.2. fom2.1 -> create index1 -> EEXIST
5.1. fom1.1 -> SUCCESS -> fom1
5.2. fom2.1 -> EEXIST -> fom2
6.1. fom1 -> lookup index1 -> SUCCESS
6.2. fom2 -> send EEXIST
7.   fom1 -> insert key1 into index1 -> send SUCCESS

Solution:
now EEXIST is treated as success.

Change-Id: Id4e91762558411855fa8fa069321b95c08131105