xcode/m0protocol: loosen runtime requirements
Since m0protocol is executed during build it should be able to run in
the environment where normal libmero initialization may not work. It's
not required to do full m0_init() just for printing xcode format,
m0_xcode_init() is sufficient in such case and it doesn't have any
runtime requirements in contrast to m0_init().

Change-Id: I081c795119b84202cb848b33d0c56e20c1bdd5fc