MERO-2337 cas/client: suppress -ENOENT error message for NEXT queries
For now CROW flag is commonly used for PUT operations. In
this case indices are physically created only on the nodes
where keys are presented. But NEXT queries are sent to all
the nodes, so some of them may return -ENOENT (index does
not exist). It is not critical, suppress these errors not
to irritate the user.

Change-Id: Ic58e1835491c9722b9682a19c58917fd45413095