MERO-2608: fix EEXIST race at m0_cob_name_add()
The check for EEXIST and the actual btree insert were not
protected with a lock. The insertion itself is protected
and it can return EEXIST error, but status of the insertion
was not checked. As result, two clients creating the same file
in parallel with O_EXCL could end up with successful result.

Now, we don't call separate lookup for EEXIST check at all.
Instead, we rely on the btree insertion result.

Reviewed-by: Maxim Medved <max.medved@seagate.com>

Change-Id: Ia271ad2074de6a46280e669bed5a165c138f989c