clovis: STABLE completion of ops when server responds and API for op return code
An operation launched through clovis will always complete with STABLE status
in case of a response from server.
A new member added to clovis op structure, op_rc, to store the return / error
code received in the reponse from the server. A new API added for applications
to access this return code.

Change-Id: Ie4637a6bb3df1f2315d4b85b53123807e4aa1c4e