MERO-2604: st/m0t1fs_client: file_creation_test() fixed
In commit 436062d file_creation_test() was improved a bit,
but with improvement a minor problem was introduced also
which caused the test mount directory to be left dirty after
the test. As result, it affected further tests and caused
them to fail (like file_creation_test one).

During the file_creation_test() we re-mount-ed m0t1fs
several times without specifying different fid_start on
each new mount. In non-oostore mode it caused FIDs re-usage
for the new files (starting from M0_MDSERVICE_START_FID).
As result, server tried to hard-link the new files to the
old ones (see m0_md_create()), but nlink was not updated
for the files (as it is usually updated by the client side
on explicit hard-linking) and hence meta-data was corrupted.
As result, the testing mount directory could not be cleaned
up properly was left dirty.

Now we just avoid umount/mount at file_creation_test() and
don't bother about hard-links as we don't support it ATM.

Change-Id: Ie7182808d2bc8416348c9c8d16b54ff5e39dd66b