MERO-2548: 12sns-cc-io-tests and 13sns-repair-quiesce ST fails on Jenkins randomly.
Aggregation group in sns repair uses accumulator copy packets to
accumulate the recovered data/parity unit in a parity group thus
recovering the lost data/parity unit. An accumulator copy packet is
launched and marked active once all the alive data/parity units required
to recover the lost units are transformed in it.Typically an aggregation
group is finalised once accumulator copy packet has completed writing
recovered data but in case of failures, not all the alive units may be
available and transformed. In such a case an accumulator copy packet may
not be launched for writing and may not be active. This is handled
accordingly during aggregation group finalisation.

Problem:
In the present situation an accumulator copy packet is marked active
after it has been launched (copy packet fom submitted to request
handler). Somehow the launched accumulator copy packet reaches its
finalisation without being marked active which leads to corresponding
aggregation group not getting finalised. Thus not allowing sns repair
operation to complete.
Speculation is that the accumulator copy packet is launched in context
of a different copy packet fom and set active. But accumulator copy
packet having its own fom is executed in a separate reqh locality.
Marking an accumulator copy packet as active (after launching) and it
reaching its finalisation, both events may happen in parallel on
different cpus, could cache coherency problems or race lead to this
issue
Fix:

Currently marking an accumulator copy packet as active just before
launching it but after meeting the relevant conditions seems to help fix
the issue as the tests pass successfully even after several runs.

MERO-2572: Mero panic:(sag->sag_base.cag_cp_local_nr == sag->sag_cp_created_nr)
at __group_next()

This assertion may not hold true always and fails in at least 2 cases as
follows:
- Concurrent delete during sns repair: In this case a cob may be deleted
while sns repair is processing a parity group and was about to create a
copy packet for a unit residing on the deleted cob, but as cob is not
found later, sns repair skips that unit and moves ahead. But due to
earlier calculation the skipped unit was expected to be available as a
local unit for that sns repair node. Due to this situation the local
unit is accounted but its corresponding copy packet is not created, this
fails the above assertion.
- In case of rebalance: In a particular situation during rebalance, in
case of multiple failures, an sns rebalance node may find a local spare
unit with data. In rebalance the data from spare is written as is and is
not required to be transformed unlike sns repair. Thus an incoming copy
packet with data from spare unit, running in some reqh locality has to
be moved to a different locality selected such as to handle concurrent
writes on the stob. To achieve this a copy packet is duplicated (without
copying its data), the source copy packet is finalised and the new one
is launched for write. To keep accounting consistency, the relevant
counters are updated. This may lead to this assertion failure due to
mismatch in read and updated counters.

Change-Id: I34c01ffea80f9771d454296d23f4790b122f91ce