MERO-2608: addon: disable flooding warnings
In prev. patch to the MERO-2608 we introduced warning
logs for cob insert duplications. As appeared, there
are many such logs in normal IO, because we re-create
spare cobs on each write to a parity group. (MERO-2620
was filed for this issue.)

So now we just disable the logging on WARN level
to avoid flooding syslog. But we still do the logging
on CALL level.

Reviewed-by: Maxim Medved <max.medved@seagate.com>

Change-Id: I50ead15765464136dacba1ccbbd84f6ab18c08bf