MERO-2649: Fail repair if no device is in REPAIRING state
Problem: During cluster testing, it was observed that a drive which was being
rebalanced, failed due to a device attach problem. Halon marked the device
as "FAILED" and restarted the repair on the drive which was already repaired.
This caused the SNS repair to hang. Ideally, the failed device should have
been marked as "REPAIRED" by Halon and if the failed device was replaced, and a
rebalance should have been re-triggered by Halon.

Root cause:
Currently, Mero SNS does not handle this inconsistency resulting from incorrect
Halon events/triggers. SNS is trying to repair a already repaired drive, due to an
incorrect Halon repair trigger. The device however, is in REBALANCING state,
and as a result, there are only outgoing groups from other SSUs, and no incoming
groups, which caused SNS repair to hang. Mero does not allow a transition from
REBALANCING to REPAIRING because disk was already repaired.

Fix:
- Check for device state as M0_PNDS_SNS_REBALANCING for the case where the
repaired device fails during rebalance, and a new repair is triggered.
If the above condition is true, the cob is not marked as failed and the
SNS iterator does not take into account this cob while calculating number of local
units (and in turn no outgoing groups are created for that failed cob).
Repair continues till all the local cobs are iterated on by the iterator
and completes (gracefully) since there are no outgoing groups.
The same checks have been added for case when a spurious rebalance event
triggers a rebalance operation when the device state is M0_PNDS_SNS_REPAIRING.

- Modified the function m0_sns_cm_is_cob_failed() and made it part of
sns operation helper vectors. The above sns operation specific checks are
now added to this function.

- Modified m0t1fs_sns_repair_1f.sh and added cases for testing the
  spurious rebalance and repair events.

Change-Id: I049d5e93495415cc2c22515eceaad0186dd06da3