rpm: correct include path in Mero pkgconfig configuration
With the current organisation of Mero headers they cannot be included in
the code as <mero/*/*.h>. A software that uses Mero headers should
explicitly specify their location with a compiler option even when
they're installed in '/usr/include/mero/'. Now a correct option is
provided in the mero.pc file:

  -I/usr/include/mero

Also, 'pkgconfig' has been added as a dependency for mero-devel rpm
package.

Change-Id: I08ff79b0bbc389deaf7603adfdbce6e8d44d2f0b