MERO-2729: fix spiel ST sporadic stucks
Sometimes Entry Point reply may return several confd
services when the configuration is updated already
on the server side. But the second confd process
was not starting until spiel client (which updates
the configuration) is finished. This caused the dead
lock, because the client tried to connect to the 2nd
confd.

Now we start 2nd confd process in parallel with the
spiel client which updates the configuration, but
with the 5 seconds delay to allow the configuration
to be updated.

Also, the FIDs on the running services on the 1st
confd process were fixed in the new configuration
to be the same as in the old one.

Revewed-by: Valery V. Vorotyntsev <valery.vorotyntsev@seagate.com>

Change-Id: Ic3cadf20d900c40b9074293648c71f4dbea4282b