scripts/halon_facts: fix default tolerance vector
The halon_facts template for minimal configuration
(single-node with parity 1) should generate minimal
tolerance vector, which is 1 disk failure only:

       - [0,0,0,0,1]

With parity 2 we can tolerate two disks already:

       - [0,0,0,0,1]
       - [0,0,0,0,2]

With 8+2 layout and 6+ nodes configuration we can
tolerate the failure of 1 node also:

       - [0,0,0,0,1]
       - [0,0,0,0,2]
       - [0,0,0,1,0]

But all this requires some scripting logic to be
implemented which is out of scope of this patch.
For now we just put the minimal tolerance vector
by default.

Reviewed-by: Nachiket Sahasrabudhe <nachiket.sahasrabuddhe@seagate.com>

Change-Id: Icd6ac656302c044acc05223f529b4958601ebbb4