be/ut/seg: suppress false warning
Move `size` assignment outside loop. It's set to the same value
unconditionally and used outside the loop. Therefore, it is more natural
to initialize it once at the beginning.

Besides, clang generate a warning that `size` can be used uninitialised.
This is not true, because `geom` contains multiple elements. Current
patch suppresses the warning.

Change-Id: Icf6a3baed1f66ab5473f0a143fbc613e909ba11a