mero.spec.in, m0confgen: Switch to python3
m0confgen fails with AssertionError when processing this input:

```bash
{
    cat <<'EOF'
[1:
{0x72| (((0x7200000000000001, 0x410)),
    [0x2:0xffffffffffffffff, 0xffffffff],
    0x8000000,
    0x5e78c800,
    0x5e78c800,
    0x5e78c800,
    "172.18.1.2@o2ib:12345:41:401",
    [0x6: (0x7300000000000001, 0x413),
        (0x7300000000000001, 0x415),
        (0x7300000000000001, 0x414),
        (0x7300000000000001, 0x412),
        (0x7300000000000001, 0x411),
        (0x7300000000000001, 0x416)])}
]
EOF
} | python2 utils/m0confgen -f xcode -t confgen
```

0xffffffffffffffff and 0xffffffff have the same type (int) in python3,
but different types (long and int) in python2.  This causes one of the
assertions to fail.

It would be easy to update (or delete) that assertion. But I prefer
to use this failure as an excuse to migrate to python3.  Maintaining
compatibility with python2 is not particularly entertaining.
And Python 2.7 will be retired in 15 months anyway, according to
https://pythonclock.org/ .

Change-Id: I9218002bc2cf278256587f0b82398a4d91a593bf