Clovis: 28clovis-sys-kvs-test failed in obj_create_double_same_id test.
This test was designed to test the case where 2 creation ops of the
same object fids are launched in parallel. One of the op is assumed
to fail. But when md_redundancy is configured larger than 1, as clovis
doesn't manage concurrent ops by design (meaning clovis doesn't hold
any object lock), there is a possibility that the 2 creation ops, each
can get both successful replies and failure replies from services and
both are considered as successful in current clovis implementation.
So the result of this test is not determined when md_redundancy is
greater than 1. This patch is to remove this test.

Change-Id: I0e3c7604a810b273635ef450831eb4df01be1f93