MERO-2833: fix panic in locality invariant
There was a race in m0_fom_block_leave() code,
between the thread was set to UNBLOCKING state and
fl_unblocking counter increment. It could happen
that the  working locality thread could not see the
incremented value yet. Which caused the panic at
invariant on:

  ergo(t->lt_state == UNBLOCKING,
     m0_atomic64_get(&loc->fl_unblocking) > 0);

Now we increment the fl_unblocking counter first
and then set the thread's state to unblocking.

Change-Id: Ib6e820af416cb6057f754652058a0c8cca4eafc3