m0crate/clovis_io: fix resources management
It was allocating and freeing object structures
(like m0_clovis_obj) on each clovis operation. So it
was not possible, for example, to create objects and
then write to the these same objects, it was needed
to open them again to be able to work with them. That
was quite inconvenient and very unoptimal in terms
of memory management and hence performance.

Now we allocate all needed objects in the beginning
of the test and free them at the end. The basic test
is passing now (at last!)

Change-Id: I123d369e40af00755f0b1dab815a9cbf341c20a3