m0crate/clovis_io: fix odd delays
We polled the value of cti_max_ops_sem semaphore to
get the idea about all Clovis asynchronous operations
completion.  And we were polling it once in a second.
This introduced a lot of bogus delays into the tests
execution, especially for IO tests with many objects.

Now we got rid of polling completely. Instead, we
down() the semaphore max_ops times (initial value).
It passes as soon as all ops are complete.

Also, we don't wait for all operations to complete
for each object in IO tests (at cr_op_io()). There
is no apparent point in this waiting.

Few minor improvements are piggybacked here also:

- added total execution time statistics for each
test (creation, write, read, deletion) and fixed the
bandwidth (BW) calculation based on this time;

- added read statistics;

- improved object ids generation randomisation and
added trace printouts for them;

- added trace printouts for the tests for the
convenience, so that we can see which thread which
test runs at the moment.

Change-Id: I1aecdd0653929318a3be800984f2a613cb5d6293