MERO-2823 fix disks messages flooding in syslog
During `hctl mero stop` or when m0d is restarted, halon notifies all
the nodes with offline message for this node as well as configuration
objects under it (which includes disks).

Now disabling the logging of offline and online messages of disk
in m0_poolmach_event_list_dump_locked(), will reduce the log
substantially.

Change-Id: I600f7246ba37dc82073b2055a236b73b66e4d5f7