be/alloc: fix a crash during large allocation
It was possible to get from a free lists a chunk with a size that is
equal to the requested size, but the chunk might have been not aligned
even to the minimal alignment. During the chunk split it is ensured that
the resulting chunk will be aligned with the right alignment, but this
may lead to having less total chunk size than was requested. The
allocator code wasn't ready for this so it panicked.

Example:
0x400100001 chunk memory start, chunk memory size = 0x1000
0x400101001 chunk memory end, next be_alloc_chunk start.

If we have an allocation request for a size 0x1000, then the chunk
above is a perfect hit.  But then the chunk memory gets aligned and
it starts looking like the following:

0x400100008 chunk memory start, chunk memory size = 0xff9
0x400101001 chunk memory end, next be_alloc_chunk start.

So now the chunk memory size is 0xff9, so we can't use it for the
allocation with size 0x1000.

The solution is to align all the chunks to the minimal alignment,
so there is always enough space in the chunk even after the final
alignment.

Change-Id: I53766b41d7db352e9db5a8cb27859e3fe6209cc3