lib/buf: add m0_buf_cmp()
This function is an analogue of memcmp(3), but also works on buffers
with different sizes.

It is common practice to represent strings with m0_buf as buffers
without trailing '\0' (See M0_BUF_INITS, m0_buf_streq). Therefore,
we can't use strcmp(3). m0_buf_cmp() can be useful in sorting m0_buf
strings in lexicographical order.

Since m0_buf may contain '\0' bytes, m0_buf_cmp() is unified
respectively. In the special case, when one buffer is a prefix for the
second, difference of buffers sizes is returned. This is because of the
situation when the first byte of the suffix is '\0'. Comparing it to
0 would return 0, but the buffers are not equal.

Change-Id: I263da38f28b38e403f47a4f02182a8f54719e0cc