mero/setup: fix buffer over-read
`ep_len' may be less than length of the endpoint. Therefore, strncpy()
will not write terminating null byte in this case and it leads to buffer
over-read in further strtok_r() call.

As solution, terminate resulting string when it's shorten.

Also, calculate `ep_len' after checking that ep is not NULL in order to
fix cppcheck warning.

Change-Id: Iaff206b5d000db6f72f3eec543233863abd77435