clovis/st/mt/clovis_mt_fom: fix buffer overflow
Key and value buffers have size of CMT_KV_SZ. Filling and checking
functions use sprintf() to write exactly CMT_KV_SZ symbols to the
buffers, but the thing is that sprintf() always writes additional
trailing '\0'. Therefore, it leads to buffer overflow.

As a solution, write (CMT_KV_SZ-1) symbols to the buffers, so 1 byte is
left for the '\0'.

Reviewed-by: Anatoliy Bilenko <anatoliy.bilenko@seagate.com>

Change-Id: I9db4abe70fc1be4648ad4ebcb81cb86df5c571cf