MERO-2881: pool: fix deadlock on confd/RM restart
Locality was blocked by the active_rm_ctx_create() call
(which tries to create the new context for the new RM)
because it requires configuration to be available
(it calls m0_conf_service_get()). But the rconfc startup
was not completed by this time yet and it could not
because the locality was blocked. Hence the deadlock.

Now we call active_rm_ctx_create() only when rconfc is
ready already, at m0_pools_common_conf_ready_async_cb().

Reviewed-by: Maxim Medved <max.medved@seagate.com>

Here is the blocked locality stack for the reference:

3  m0_semaphore_timeddown (semaphore, abs_timeout) at lib/user_space/semaphore.c:124
4  m0_chan_timedwait (link, abs_timeout) at lib/chan.c:354
5  m0_sm_timedwait (mach, states=states@entry=2112, deadline) at sm/sm.c:395
6  rconfc_gate_check (confc) at conf/rconfc.c:2113
7  m0_confc_ctx_init (ctx, confc) at conf/confc.c:644
8  sm_waiter_init (w, confc) at conf/confc.c:791
9  m0_confc__open_sync (result, origin, path) at conf/confc.c:895
10 confc_obj_get (confc, fid, result) at conf/helpers.c:43
11 m0_conf_service_get (confc, fid, out) at conf/helpers.c:62
12 active_rm_ctx_create (pc, service_connect) at pool/pool.c:909
13 service_ctx_ha_entrypoint_cb (clink) at pool/pool.c:1005
14 clink_signal (clink) at lib/chan.c:133
15 chan_signal_nr (chan, nr=0) at lib/chan.c:152
16 m0_chan_broadcast (chan) at lib/chan.c:172
17 state_set (mach, state=-1, state@entry=10, rc=0) at sm/sm.c:450
18 m0_sm_state_set (mach, state=10) at sm/sm.c:469
19 ha_entrypoint_client_fom_fini (fom) at ha/entrypoint.c:680
20 fom_exec (fom) at fop/fom.c:757
21 loc_handler_thread (th) at fop/fom.c:873

Change-Id: I775a2c1b30c8c9d23639be1db06bc67535d58958