MERO-2877 Fix a panic in entry point callback during active rm restart
In multiple confd + rm setup, after active rm is restarted,
new entrypoint is requested. Then in service_ctx_ha_entrypoint_cb()
while comparing with the rm fid from entrypoint reply, old rm ctx is
referred. When it is already cleaned up, it causes a panic.

This is fixed by checking pc->pc_rm_ctx is NULL or not. If it is NULL
then new context is directly created for the rm from entrpoint reply.

Reviewed-by: Andriy Tkachuk <andriy.tkachuk@seagate.com>

Change-Id: I8a3a00a995fcf744798f63d0766e55f7e14b1aa7