Replace PAGE_CACHE_SIZE with PAGE_SIZE
PAGE_CACHE_SIZE always was defined as PAGE_SIZE and only was planned to
be something different.

From the description of the macros:
The page cache can done in larger chunks than one page ... Or rather, it
_will_ be done in larger chunks.

But in Linux 4.6 the PAGE_CACHE_ macros were removed at all. Follow the
reference link for more details.

Since PAGE_CACHE_SIZE macro always was just a synonym for PAGE_SIZE, it
is safe to replace it for all kernel versions. It will increase
portability of Mero kernel code.

Do the same for PAGE_CACHE_SHIFT and PAGE_CACHE_MASK.

Change-Id: Ib0d6b3c48c7a674b623aa1f2632bae78d6a14776
Reference: https://lkml.org/lkml/2016/3/21/274