lib/ut/memory: fix poisoning check on some systems
libm0-ut:memory test checks if memory is poisoned after free. But since
free(3) uses beginning of the memory buffer for its purposes, the test
checks (ptr+8) rather than `ptr'.

free(3) implementation is specific for a particular system/libc.
On a system with different glibc it can modify first 16 bytes instead of
8, as the test assumes. Therefore, the test fails.

As solution, check if (ptr+16) is poisoned. It will be more portable.

Also avoid void* in calculation and cast `ptr' to char*.

Change-Id: I19ce65eea5b03bbfa9bf555a46213741a313c5a2