MERO-2892: fix client crash on SSU restart
The issue reproduced when some new files were written
while the SSU was down. It was caused by the stale
pointer to the currect pool version (pc_cur_pver) at
m0_pools_common_conf_ready_async_cb() which was referenced.

When some new file is written while SSU is down the new
pool version is created. But when the SSU is restarted
(followed by the configuration update) - the pool version
is no longer available in the new configuration conf cache,
so it is freed at pools_common_pools_update_or_cleanup().
But the pc_cur_pver pointer was not reset.

Now we just drop the pc_cur_pver pointer completely
as it is not needed anymore - the pool version is selected
for each new file/object on its creation. (We still left
the pointer at the m0_pools_common structure though because
the Clovis UTs are using it heavily, but this should be
fixed later.)

Reviewed-by: Rajanikant Chirmade <rajanikant.chirmade@seagate.com>

Change-Id: Ie6f61694af48ea01198b8b3424b9ecfc6ce16fd3