lib/assert: introduce m0_panic_only()
m0_trace_args_unpack() calls m0_arch_panic(). But there is an issue
with the va_list argument. Currently, m0_trace_args_unpack() passes
NULL there and it fails on an architecture where va_list is not a
pointer. We could define a fake va_list variable to pass it where it is
not needed, but there is no portable initializer for va_list. Therefore,
introduce new interface m0_panic_only() which accepts ellipsis instead
of va_list.

Change-Id: Ie307be57c21c3d5d7db33ea061b0e022dfad7b23